problem: code duplication for tags in forms #518
solution: Adds a tagInput component to reuse the code & reduce code duplication in the forms
