+++ imports/ui/pages/community/socialResourceForm.html
@@ -32,14 +32,8 @@
                       </div>
                       <small><i id="description-chars" class="text-muted"></i></small>
 
-                      <div class="group-title">Tags</div>
-                      <div class="input-group">
-                        <select multiple="true" name="tags[]" id="tags" class="form-control select2">
-                          {{#each tags}}
-                          <option value="{{name}}" {{isTagIncluded project.tags _id}}>{{name}}</option>
-                          {{/each}}
-                        </select>
-                      </div>
+                      {{> tagInput values=socialTags}}
+
                       <div class="input-group">
                             {{> reCAPTCHA}}
                             <div id="captchaError" class="invalid-feedback"></div>
