+++ imports/ui/pages/community/socialResourceForm.js
@@ -27,14 +27,6 @@ Template.socialResourceFormTemp.onCreated(function() {
 })
 
 Template.socialResourceFormTemp.onRendered(function() {
-    this.autorun(() => {
-        let tags = (socialResources.findOne({
-            _id: FlowRouter.getParam('id')
-        }) || {}).tags || []
-
-        $('#tags').val(tags.map(i => i.name))
-        $('#tags').trigger('change')
-    })
 
     $('#tags').select2({
         tags: true,
@@ -48,17 +40,7 @@ Template.socialResourceFormTemp.onRendered(function() {
 Template.socialResourceFormTemp.helpers({
     add: () => FlowRouter.current().route.name === 'editSocialResource' ? false : true,
     Resource: () => {return socialResources.findOne({ _id: FlowRouter.getParam('id') })},
-    tags: () => { 
-
-        let tags = Array.from(Tags.find({
-            name: {
-                $not: new RegExp('built-(for|on)-cardano', 'i') // dont include these tags
-            }
-        }))
-
-        tags = _.uniqBy(tags, 'name');
-        return tags
-    }
+    socialTags: () => (socialResources.findOne({ _id: FlowRouter.getParam('id') }) || {}).tags || []
 })
 
 Template.socialResourceFormTemp.events({
