+++ imports/ui/pages/learn/learnForm.html
@@ -32,14 +32,7 @@
             <input type="radio" name="difficultyLevel" value="advanced"> Advanced
             <div id="typeError" class="invalid-feedback"></div>
           </div>
-          <div class="group-title">Tags</div>
-          <div class="input-group">
-            <select multiple="true" name="tags[]" id="tags" class="form-control select2">
-              {{#each tags}}
-              <option value="{{name}}" {{isTagIncluded project.tags _id}}>{{name}}</option>
-              {{/each}}
-            </select>
-          </div>
+          {{> tagInput values=learnTags}}
           <div class="input-group">
               {{> reCAPTCHA}}
               <div id="captchaError" class="invalid-feedback"></div>
