+++ imports/ui/pages/projects/projects.scss
@@ -1,4 +1,5 @@
 .project-form{
+  
   span.instructions {
     display: none;
     align-items: center;
@@ -8,6 +9,7 @@
       padding-right: 0.5em;
     }
   }
+
   #description:focus + span.instructions {
     display: flex !important;
   }
